LayoutTests/fast/js/dfg-cfg-simplify-redundant-dead-get-local-expected.txt

Tests if the CFG simplifier gracefully handles the case where Block #1 and #2 are merged, #1 has a dead GetLocal, and #2 has a live GetLocal on the same local.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS array = []; calcError(); array[0] is 2.23606797749979 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
